Enable Offline Token Rotation
To make life easier for clients that consume service account tokens,
rather than manually monitoring for expiry and rotating them, we want an
automated approach to this.  With the current Kubernetes secret based
architecture, we will see a minimal downtime window of upto 2 minutes
every 60 days, which is 99.998% uptime at worst, which isn't horrible.